Move vscode-message-manager.ts in its own sub-directory
