+++ vscode-trace-common/src/messages/vscode-message-manager.ts
