+++ vscode-trace-webviews/src/trace-explorer/available-views/vscode-trace-explorer-views-widget.tsx
@@ -7,7 +7,7 @@ import { ReactAvailableViewsWidget } from 'traceviewer-react-components/lib/trac
 import 'traceviewer-react-components/style/trace-explorer.css';
 import { Experiment } from 'tsp-typescript-client/lib/models/experiment';
 import { TspClientProvider } from '../../common/tsp-client-provider-impl';
-import { VsCodeMessageManager, VSCODE_MESSAGES } from 'vscode-trace-common/lib/vscode-message-manager';
+import { VsCodeMessageManager, VSCODE_MESSAGES } from 'vscode-trace-common/lib/messages/vscode-message-manager';
 import '../../style/react-contextify.css';
 import '../../style/trace-viewer.css';
 import JSONBigConfig from 'json-bigint';
