+++ vscode-trace-webviews/src/trace-viewer/vscode-trace-viewer-container.tsx
@@ -13,7 +13,7 @@ import 'traceviewer-react-components/style/trace-context-style.css';
 import { Experiment } from 'tsp-typescript-client/lib/models/experiment';
 import { OutputDescriptor } from 'tsp-typescript-client/lib/models/output-descriptor';
 import { TspClientProvider } from '../common/tsp-client-provider-impl';
-import { VsCodeMessageManager, VSCODE_MESSAGES } from 'vscode-trace-common/lib/vscode-message-manager';
+import { VsCodeMessageManager, VSCODE_MESSAGES } from 'vscode-trace-common/lib/messages/vscode-message-manager';
 import { convertSignalExperiment } from '../common/vscode-signal-converter';
 import '../style/trace-viewer.css';
